pyproject.toml
[build-system]
requires = ["setuptools>=77.0.3"]
build-backend = "setuptools.build_meta"

[project]
name = "omega-research"
version = "0.6.0"
description = "Standalone Python-first runtime and CLI surface for the OMEGA open-mathematics research lab."
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
authors = [
  { name = "MicroPhoenix open-source community" },
]
dependencies = [
  "PyYAML==6.0.1",
  "jsonschema==4.23.0",
]
keywords = ["mathematics", "research-software", "lean4", "provenance", "experiment-tracking"]
classifiers = [
  "Private :: Do Not Upload",
  "Development Status :: 4 - Beta",
  "Intended Audience :: Science/Research",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Scientific/Engineering :: Mathematics",
]

[project.optional-dependencies]
smt = ["z3-solver>=4.12"]
sat = ["python-sat>=1.8"]
cas = ["sympy>=1.13"]
research = ["numpy>=1.26", "scipy>=1.13", "mpmath>=1.3", "matplotlib>=3.8"]
all = ["z3-solver>=4.12", "python-sat>=1.8", "sympy>=1.13", "numpy>=1.26", "scipy>=1.13", "mpmath>=1.3", "matplotlib>=3.8"]

[project.scripts]
omega-export-runtime-baseline = "export_runtime_baseline:main"
omega-export-execution-maturity = "export_execution_maturity:main"
omega-export-ft-scaffold-gate = "export_ft_scaffold_gate:main"
omega-runner = "omega_runner:main"
omega-generate-index = "generate_index:main"
omega-generate-experiment-index = "generate_experiment_index:main"
omega-import-einstein-arena = "import_einstein_arena:main"
omega-scaffold-problem = "scaffold_problem:main"
omega-validate-registry = "validate_registry:main"
omega-query = "experiment_query:main"
omega-literature = "literature_adapter:main"
omega-lean = "lean_adapter:main"
omega-solve = "solver_adapter:main"
omega-cas = "cas_adapter:main"
omega-workflow = "omega_workflow:main"
omega-orchestrate = "agent_orchestrator:main"
omega-verify-evidence = "verify_evidence:main"
omega-model-router = "model_router:main"
omega-einstein-arena = "einstein_arena_adapter:main"
omega-leancop-bridge = "leancop_bridge:main"
omega-proof-repair = "proof_repair_loop:main"
omega-verify-version-sync = "verify_version_sync:main"

[tool.setuptools]
py-modules = [
  "export_runtime_baseline",
  "export_execution_maturity",
  "export_ft_scaffold_gate",
  "omega_runner",
  "generate_index",
  "generate_experiment_index",
  "import_einstein_arena",
  "scaffold_problem",
  "validate_registry",
  "experiment_query",
  "literature_adapter",
  "lean_adapter",
  "solver_adapter",
  "cas_adapter",
  "omega_workflow",
  "agent_orchestrator",
  "verify_evidence",
  "model_router",
  "einstein_arena_adapter",
  "leancop_bridge",
  "proof_repair_loop",
  "verify_version_sync",
]

[tool.setuptools.package-dir]
"" = "scripts"